Upgrade `perchance` to v0.4.0
Removes dependency on `macaw` and `glam` which got quite messy as that makes `ark` use old versions of its own dependencies from the repo.

This is still the case now with `perchance` which is also not ideal, but at least the impact of just `perchance` is much smaller than also having all of `glam` and `macaw` duplicated.